requirements (2).txt
streamlit==1.32.0
python-docx==0.8.11
lxml==4.9.2
python-dateutil==2.9.0.post0
Pillow==10.0.1